+++ database/migrations/2025_09_14_102950_create_orders_table.php
@@ -10,13 +10,13 @@ class CreateOrdersTable extends Migration {
 		Schema::create('orders', function(Blueprint $table) {
 			$table->increments('id');
 			$table->timestamps();
-			$table->integer('client_id')->unsigned();
+			$table->bigInteger('client_id')->unsigned();
 			$table->tinyInteger('status');
 			$table->tinyInteger('payment_method');
 			$table->decimal('price', 10,2);
 			$table->decimal('shipping_cost', 10,2);
 			$table->decimal('total_price', 10,2);
-			$table->integer('sale_id')->unsigned()->nullable();
+			$table->bigInteger('sale_id')->unsigned()->nullable();
 		});
 	}
