+++ database/migrations/2025_09_14_102950_create_returns_table.php
@@ -10,9 +10,9 @@ class CreateReturnsTable extends Migration {
 		Schema::create('returns', function(Blueprint $table) {
 			$table->increments('id');
 			$table->timestamps();
-			$table->integer('sale_id')->unsigned();
-			$table->integer('safe_id')->unsigned();
-			$table->integer('user_Id')->unsigned();
+			$table->bigInteger('sale_id')->unsigned();
+			$table->bigInteger('safe_id')->unsigned();
+			$table->bigInteger('user_Id')->unsigned();
 			$table->string('return_number')->nullable();
 			$table->decimal('return_amount', 10,2);
 			$table->text('reason')->nullable();
